+++ src/main/java/org/datadog/jmxfetch/JmxComplexAttribute.java
@@ -76,6 +76,9 @@ public class JmxComplexAttribute extends JmxSubAttribute {
             this.subAttributeList.addAll(JeeStatisticsAttributes.attributesFor(attributeValue));
         } else if (JeeStatisticsAttributes.isJeeStat(attributeValue)) {
             this.subAttributeList.addAll(JeeStatisticsAttributes.getStatisticNames(attributeValue));
+        } else {
+            log.trace("beanName {} attributeValue type {}: no match {}",
+                getBeanName(), attributeValue.getClass(), attributeValue);
         }
     }
 
@@ -134,21 +137,6 @@ public class JmxComplexAttribute extends JmxSubAttribute {
         }
 
         return matchAttribute(configuration) && !excludeMatchAttribute(configuration);
-    }
-
-    private boolean matchSubAttribute(
-            Filter params, String subAttributeName, boolean matchOnEmpty) {
-        if ((params.getAttribute() instanceof Map<?, ?>)
-                && ((Map<String, Object>) (params.getAttribute()))
-                        .containsKey(subAttributeName)) {
-            return true;
-        } else if ((params.getAttribute() instanceof List<?>
-                && ((List<String>) (params.getAttribute())).contains(subAttributeName))) {
-            return true;
-        } else if (params.getAttribute() == null) {
-            return matchOnEmpty;
-        }
-        return false;
     }
 
     private boolean matchAttribute(Configuration configuration) {